Add conditional dark theme styles depending on image/video colors and selected overlay colors
