+++ packages/block-library/package.json
@@ -38,6 +38,7 @@
 		"@wordpress/keycodes": "file:../keycodes",
 		"@wordpress/viewport": "file:../viewport",
 		"classnames": "^2.2.5",
+		"fast-average-color": "4.3.0",
 		"lodash": "^4.17.11",
 		"memize": "^1.0.5",
 		"url": "^0.11.0"
